Topology: Fix undefined behaviour on implicit conversions

+++ liblwgeom/liblwgeom_topo.h
@@ -222,7 +222,7 @@ typedef struct LWT_BE_CALLBACKS_T {
 					   double dist,
 					   uint64_t *numelems,
 					   int fields,
-					   int limit);
+					   int64_t limit);
 
   /**
    * Insert nodes
@@ -282,7 +282,7 @@ typedef struct LWT_BE_CALLBACKS_T {
 					   double dist,
 					   uint64_t *numelems,
 					   int fields,
-					   int limit);
+					   int64_t limit);
 
   /**
    * Get next available edge identifier
